feat(engine): migrate event sub process activation

+++ engine/src/main/java/io/zeebe/engine/processing/common/EventTriggerBehavior.java
@@ -29,7 +29,6 @@ import io.zeebe.engine.state.mutable.MutableZeebeState;
 import io.zeebe.protocol.impl.record.value.processinstance.ProcessInstanceRecord;
 import io.zeebe.protocol.record.intent.ProcessInstanceIntent;
 import io.zeebe.protocol.record.value.BpmnElementType;
-import java.util.function.ToLongFunction;
 import java.util.stream.Collectors;
 import org.agrona.DirectBuffer;
 
@@ -104,7 +103,8 @@ public class EventTriggerBehavior {
   public void triggerEventSubProcess(
       final ExecutableStartEvent startEvent,
       final long flowScopeElementInstanceKey,
-      final ProcessInstanceRecord recordValue) {
+      final ProcessInstanceRecord recordValue,
+      final DirectBuffer variables) {
 
     final var flowScopeElementInstance =
         elementInstanceState.getInstance(flowScopeElementInstanceKey);
@@ -124,41 +124,23 @@ public class EventTriggerBehavior {
                 flowScopeElementInstance.getValue(),
                 flowScopeElementInstance.getState());
 
-    triggerEvent(
-        flowScopeContext,
-        eventTrigger -> {
-          final var eventSubProcessElementId = startEvent.getEventSubProcess();
-          final var record =
-              getEventRecord(recordValue, eventTrigger, BpmnElementType.EVENT_SUB_PROCESS)
-                  .setFlowScopeKey(flowScopeElementInstanceKey)
-                  .setElementId(eventSubProcessElementId);
-
-          if (startEvent.interrupting()) {
-            unsubscribeFromEvents(flowScopeContext);
-            triggerInterruptingEventSubProcess(flowScopeContext, record);
-          } else {
-            eventScopeInstanceState.deleteTrigger(
-                flowScopeContext.getElementInstanceKey(), eventTrigger.getEventKey());
-            // activate non-interrupting event sub-process
-            publishActivatingEvent(record);
-          }
-          // the key is used to set temporary variables on
-          // we moved that logic for event sub process already, but not for boundary events
-          // we still need to return a key until they are migrated; -1 is ignored.
-          return -1L;
-        });
-  }
-
-  private void triggerInterruptingEventSubProcess(
-      final BpmnElementContext flowScopeContext, final ProcessInstanceRecord eventRecord) {
-
-    unsubscribeFromEvents(flowScopeContext);
-
-    final var noActiveChildInstances = terminateChildInstances(flowScopeContext);
-    if (noActiveChildInstances) {
-      // activate interrupting event sub-process
-      publishActivatingEvent(eventRecord);
+    final var eventTrigger = eventScopeInstanceState.peekEventTrigger(flowScopeElementInstanceKey);
+    if (eventTrigger == null) {
+      // the activity (i.e. its event scope) is left - discard the event
+      return;
     }
+
+    if (startEvent.interrupting()) {
+      unsubscribeFromEvents(flowScopeContext);
+
+      final var noActiveChildInstances = terminateChildInstances(flowScopeContext);
+      if (!noActiveChildInstances) {
+        // activation of event sub process happens in flow scope of last child terminated
+        return;
+      }
+    }
+
+    activateTriggeredEvent(startEvent, flowScopeElementInstanceKey, recordValue, variables);
   }
 
   private boolean terminateChildInstances(final BpmnElementContext flowScopeContext) {
@@ -184,7 +166,8 @@ public class EventTriggerBehavior {
               childInstanceContext.getRecordValue());
         }
 
-      } else if (childInstanceContext.getIntent() == ProcessInstanceIntent.ELEMENT_COMPLETED) {
+      } else if (!MigratedStreamProcessors.isMigrated(childInstanceContext.getBpmnElementType())
+          && childInstanceContext.getIntent() == ProcessInstanceIntent.ELEMENT_COMPLETED) {
         // clean up the state because the completed event will not be processed
         eventScopeInstanceState.deleteInstance(childInstanceContext.getElementInstanceKey());
         elementInstanceState.removeInstance(childInstanceContext.getElementInstanceKey());
@@ -296,26 +279,8 @@ public class EventTriggerBehavior {
     }
 
     if (flowScope.getElementType() == BpmnElementType.EVENT_SUB_PROCESS) {
-
-      // first activate the event sub process
-      eventRecord
-          .setBpmnElementType(BpmnElementType.EVENT_SUB_PROCESS)
-          .setElementId(flowScope.getId());
-
-      final var eventSubProcessKey = keyGenerator.nextKey();
-      stateWriter.appendFollowUpEvent(
-          eventSubProcessKey, ProcessInstanceIntent.ELEMENT_ACTIVATING, eventRecord);
-
-      // event sub process is not yet migrated
-      // todo(#6196): on migration of event sub process we want immediately move activated for the
-      // event sub process
-      //      stateWriter.appendFollowUpEvent(
-      //          eventSubProcessKey, ProcessInstanceIntent.ELEMENT_ACTIVATED, eventRecord);
-      //      eventRecord.setFlowScopeKey(eventSubProcessKey);
-
-      // event sub process will activate start event
-      return; // todo(#6196) remove this line after migrating event sub process
-      // now we can trigger the start event
+      activateEventSubProcess(triggeredEvent, flowScope);
+      return;
     }
 
     eventRecord
@@ -341,5 +306,26 @@ public class EventTriggerBehavior {
 
     commandWriter.appendFollowUpCommand(
         eventInstanceKey, ProcessInstanceIntent.COMPLETE_ELEMENT, eventRecord);
+  }
+
+  private void activateEventSubProcess(
+      final ExecutableFlowElement triggeredEvent, final ExecutableFlowElement flowScope) {
+    // first activate the event sub process
+    eventRecord
+        .setBpmnElementType(BpmnElementType.EVENT_SUB_PROCESS)
+        .setElementId(flowScope.getId());
+
+    final var eventSubProcessKey = keyGenerator.nextKey();
+    stateWriter.appendFollowUpEvent(
+        eventSubProcessKey, ProcessInstanceIntent.ELEMENT_ACTIVATING, eventRecord);
+    stateWriter.appendFollowUpEvent(
+        eventSubProcessKey, ProcessInstanceIntent.ELEMENT_ACTIVATED, eventRecord);
+
+    eventRecord
+        .setFlowScopeKey(eventSubProcessKey)
+        .setBpmnElementType(triggeredEvent.getElementType())
+        .setElementId(triggeredEvent.getId());
+
+    commandWriter.appendNewCommand(ProcessInstanceIntent.ACTIVATE_ELEMENT, eventRecord);
   }
 }
